move auth to dependencies
