+++ app/api/v1/endpoints/user.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends
 
-from app.core.auth import get_current_user
+from app.dependencies.auth import get_current_user
 from app.dependencies.services import get_user_service
 from app.models.user import User
 from app.schemas.common import BaseResponse
